Only fetch packages for current OS
This way it only gets a third of the packages. It doesn't really have a
noticeable performance improvement, but at least it makes it easier to
work with.

There not much reason for allowing packages from other OSs except for
Linux for use with WINE, but it's not necessary right now.