Differentiate between official releases and others
The official releases are meant to be more of a "pinned" group of
packages that the user installs to have a specific version without
any update checking. It was conflicting with the same packages that
where sourced from other places like LTS and Stable, which are
subject to be checked for updates.

To achieve this I had to add an "-official" suffix to the name of the
Official Release packages and download them into the 'temp_dir',
renaming them afterwards into 'packages_dir'. This prevents the user
from using an actual /tmp directory due to it having a different
filesystem. Having it downloaded and extracted inside ~/.cache is fine
too, since moving it doesn't cost anything.