Refactor Settings
I was using the 'config' crate but I didn't really need it and it only
made code awkward. Now I'm working directly on the Settings
struct and can benefit from type checking, eliminating lots of possible
developer errors.

I also made the decision to lock the configuration file into a binary so the
user can only change it through the provided interfaces. This eliminates
a significant amount of error handling that would be required otherwise.

One other thing it gives me is the ability to expand what I store in
Settings. I couldn't store the time of the previous update in the config
file because users may mess with that, but now there's no issue.

Now, I know that a motivated enough user can still change it, especially
since it's not actually encrypted, but those people are welcome to all
the panics they want. I'm not to blame for their tinkering. It would be
like blaming a game developer for not caring about the bugs Cheat Engine
users get when cheating in their single-player game.