Hardcode entry count for official packages
I was checking all archives so I thought I may as well get all their
"lengths". Now the progress bars for older packages are natural-looking.
And any newer packages use 4800, which is around 150 higher than the
largest count.

If there was an actual .len() method for tar archives I could get rid of
this and expose the entry count to the user instead of using
percentages, since they sometimes get stuck at 100% for a few seconds
due to rounding.