Add faulty install functionality to GUI
It works for installing packages except for updates. Can't figure
out why. The packages marked as updates also interfere with the same
packages in their respective tabs, not showing any progress (even though
they are actually being installed) if you visit the Updates tab first.

There are also other issues with the install progress specifically,
like a panic when extracting older packages, and the inability to
display an extraction progress bar for Linux due to the way the relevant
crates deal with data internally, not allowing to pass it around inside the
stream.

This is a commit to save my progress before I do a major refactor. I
really dislike the amount of repetition the code has now for each tab.