Fix opening packages on Linux (and cleanup)
It seems Blender's output can't be piped out into nothingness since it
prevents Blender from opening its GUI, even though it actually launched.

It's not a big deal since by the point Blender is opening, the launcher
is done, and the user might want to see the messages if they're
launching from a terminal, but the somewhat problematic part is that
control of the terminal is given back to the user so they can input
commands even though Blender may still keep printing into it.