Add PKGBUILD
For automatically publishing to the AUR on release. The
REPLACE_RELEASE_VERSION and REPLACE_SHA256SUMS_X86_64
variables are replaced in the workflow with their respective values, so
the only reason to change this file is to update the dependencies or
switch the target to `musl`, if I ever get it to work.