Fix `iced` backend on `noveau` and `radeon`
`iced`'s default backend on Linux requires Vulkan, which neither
Nvidia's `noveau` nor AMD's `radeon` drivers support. It's safer to use
the `glow` backend as OpenGL is more compatible and none of the
limitations are a problem at the moment.

NOTE: `glow` shouldn't be used for macOS as OpenGL is deprecated there.