Add changing paths for the program's locations
Can change the locations for databases, installed packages and cache
files. Can also change the config file's location by setting the
ABLAVEMA_CONFIG_FILE environment variable to a file, which will be
created with the specified name.

A known issue is that if the packages and cache directories are set to
paths on different mount points, installation will fail at the last step
of renaming, since that doesn't work across filesystem boundries.