Shorten calls to `Settings`
Having to write something as asinine as
`SETTINGS.read().unwrap().sad_field`
every time I need to use a setting was driving me nuts.

Now, with the new style it's as easy as
`get_setting().happy_field`
and so is for the other related functionality.

Also has the benefit of not breaking code into lines as often.