Improve settings formatting
By that I mean the fact that `rustfmt` doesn't work for code blocks
which have long strings. It just doesn't want to format anything that
has them, which meant that I had to format it myself. It was getting
annoying so I decided to break each setting into its own variable and
use the method for long strings that makes them start at the very start
of the line, using the backslash to escape newlines to not go over 100
characters in width, which makes `rustfmt` work properly.

And I guess I improved some descriptions and whatever on the way.