Add Windows icon
Should've really been in the 0.1.0 release. Used Blender's main logo but
with the centre having an `a` instead of a circle. This will be easier
for users used to the normal Blender icon, since the outer shape and the
colours are the same. Can't beat that colour scheme, really.

Used a GIMP script for the `.ico` generation. It can be found here:
https://gist.github.com/YeldhamDev/45d4677fd4849b920c5855653ecce3e9
Seems to work just fine at all view levels, but I don't know whether any
optimisations can be made.

Also added some info for the "Details" tab of the "Properties" of the
executable. Don't know if I'm missing anything else. Using `winres`
instead of `embed-resource` because it handles the version change
automatically.