Fix panic on getting date from outdated archives
The correct date was taken from tha Archived packages, but if new Stable
or LTS packages were fetched without also fetching new Archived ones, it
would obviously not find it.

Now the issue is that the time of those would stay at 00:00:00 until the
new Archived packages are fetched. I could maybe force the Archived
packages to update if either Stable or LTS found a new package, but it's
questionable whether that's needed.