Fix self-updater error when filesystems differ
Happened only when the executable and the cache directory reside on
different filesystems. Noticed it after trying to use the self-updater
from the AUR package, which was installed in `/usr/bin/ablavema`, with
the cache in `~/.cache/ablavema`, my HOME residing on a separate drive.

This fix requires it to first move the new executable to the same
location as the current one as a temporary file, then rename it,
replacing the curerent executable. Needed because moving it in one go
throws a "Text file busy (os error 26)" error.