Fix default time on stable and LTS packages
This gets rid of the stinky way in which the time was looked for in the
archived packages. But it wouldn't find the new ones if they weren't
fetched like with the stable and LTS ones, so it'd still display
00:00:00 until the archived package list was also fetched.

Now, when fetching stable or LTS packages, they fetch the relevant
packages from the archive's specific release page based on the version
and get the date from that.