Remove CLI
Implementing new features or making changes to the core functionality
basically meant doing double the work, so I scraped the CLI since it's
probably not going to be used much anyway.

This also gets rid of some TODOs and dependencies so I'm quite happy to
remove it. And it will make it easier to refactor moving forward.