Fix #5: use `ron` instead of `bincode`
This, combined with `#[serde(default)]` should fix the issue of `memory
allocation of "many exabytes" failed` along with cases where I'll change
the Settings struct later on.

I haven't added `#[serde(default)]` to any of the other structs because
I'm not sure exactly how it works for heavily nested structs. I made it
so the error reporting is a bit better, so hopefully that'll help if it
happens again. Also the files are now readable, so it's easier to debug.